[#44] Refactor: calendar 폴더명 수정 >> calendar-monthly
calendar-compound 폴더와 혼동 방지 목적
